Move core->foundation, add docs
