+++ auto_submit/lib/server/request_handler.dart
@@ -7,10 +7,17 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 import 'package:shelf/shelf.dart';
 
+import '../service/config.dart';
 import 'exceptions.dart';
 
 @immutable
 abstract class RequestHandler {
+  const RequestHandler({
+    required this.config,
+  });
+
+  final Config config;
+
   /// Services an HTTP request.
   ///
   /// The default implementation will respond with HTTP 405 method not allowed.
