+++ auto_submit/lib/service/config.dart
@@ -6,7 +6,7 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:auto_submit/core/providers.dart';
+import 'package:auto_submit/foundation/providers.dart';
 import 'package:corsac_jwt/corsac_jwt.dart';
 import 'package:github/github.dart';
 import 'package:http/http.dart' as http;
@@ -34,6 +34,8 @@ class Config {
     // GitHub's secondary rate limits are run into very frequently when making auth tokens.
     final String token = await cache['githubToken'].get(
       _generateGithubToken,
+      // Tokens have a TTL of 10 minutes. AppEngine requests have a TTL of 1 minute.
+      // To ensure no expired tokens are used, set this to 10 - 1, with an extra buffer of a duplicate request.
       const Duration(minutes: 8),
     );
